back/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.57.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.27.2
environ==1.0
filelock==3.14.0
fsspec==2024.5.0
huggingface-hub==0.23.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
